requirements.txt
cmake>=3.13.2
plyfile
tensorboardX
pyyaml
scipy
six
pandas
